+++ main.h
@@ -16,7 +16,7 @@ int _putchar(char c);
 void _puts(char *str);
 int _strlen(char *s);
 int _check_strings(const char *chk_format, ...);
-int convert_binery(unsigned int dec);
+unsigned long convert_binery(unsigned int dec);
 
 /* PRINT FUNCTIONS */
 int print_character(va_list arguments, int printed_counter);
